Add MaintenanceType for determining isMaintenanceEnabled.
Intension: Centralize information inside of the library instead of having
the app do it. Added new extension function to allow the response to contain the
logic. Returns MaintenanceType Enum to give consumer more information.
Reason: Previously the apps had to do this logic and it was messy, so I centralized
it inside of an extension function. Pass in te information and the MaintenanceType
enum will be returned.
Approach: Add new extension function with Enum return type to provide more information
to the app that consumes this library.